+++ Modules/Intent.Modules.VisualStudio.Projects/Templates/CoreWeb/AppSettings/AppSettingsModel.cs
@@ -4,12 +4,13 @@ namespace Intent.Modules.VisualStudio.Projects.Templates.CoreWeb.AppSettings
 {
     public class AppSettingsModel
     {
-        public AppSettingsModel(RuntimeEnvironmentModel runtimeEnvironment, string location, bool requiresSpecifiedRole, bool includeAllowHosts = true)
+        public AppSettingsModel(RuntimeEnvironmentModel runtimeEnvironment, string location, bool requiresSpecifiedRole, bool includeAllowHosts, bool includeAspNetCoreLoggingLevel)
         {
             RuntimeEnvironment = runtimeEnvironment;
             Location = location;
             RequiresSpecifiedRole = requiresSpecifiedRole;
             IncludeAllowHosts = includeAllowHosts;
+            IncludeAspNetCoreLoggingLevel = includeAspNetCoreLoggingLevel;
 		}
 
         public RuntimeEnvironmentModel RuntimeEnvironment { get; }
@@ -19,6 +20,7 @@ namespace Intent.Modules.VisualStudio.Projects.Templates.CoreWeb.AppSettings
         public bool RequiresSpecifiedRole { get; }
 
         public bool IncludeAllowHosts { get; }
+		public bool IncludeAspNetCoreLoggingLevel { get; }
 
 	}
 }