+++ Modules/Intent.Modules.VisualStudio.Projects/Templates/CoreWeb/AppSettings/AppSettingsTemplateRegistration.cs
@@ -37,7 +37,9 @@ namespace Intent.Modules.VisualStudio.Projects.Templates.CoreWeb.AppSettings
                     Location = "",
                     RequiresSpecifiedRole = false,
                     IncludeAllowHosts = true,
-                })
+					IncludeAspNetCoreLoggingLevel = true,
+
+				})
                 .Union(_metadataManager.VisualStudio(applicationManager).GetCSharpProjectNETModels()
                     .Where(x => x.GetNETSettings().SDK().IsMicrosoftNETSdkWeb())
                     .Select(x => new
@@ -47,6 +49,7 @@ namespace Intent.Modules.VisualStudio.Projects.Templates.CoreWeb.AppSettings
                         Location = "",
                         RequiresSpecifiedRole = false,
 						IncludeAllowHosts = true,
+						IncludeAspNetCoreLoggingLevel = true,
 					}))
 				.Union(_metadataManager.VisualStudio(applicationManager).GetCSharpProjectNETModels()
 					.Where(x => x.GetNETSettings().SDK().IsMicrosoftNETSdkWorker())
@@ -57,6 +60,7 @@ namespace Intent.Modules.VisualStudio.Projects.Templates.CoreWeb.AppSettings
 						Location = "",
 						RequiresSpecifiedRole = false,
 						IncludeAllowHosts = false,
+						IncludeAspNetCoreLoggingLevel = false,
 					}))
 				.Union(_metadataManager.VisualStudio(applicationManager).GetCSharpProjectNETModels()
                     .Where(x => x.GetNETSettings().SDK().IsMicrosoftNETSdkBlazorWebAssembly())
@@ -67,17 +71,18 @@ namespace Intent.Modules.VisualStudio.Projects.Templates.CoreWeb.AppSettings
                         Location = "wwwroot",
                         RequiresSpecifiedRole = true,
 						IncludeAllowHosts = true,
+						IncludeAspNetCoreLoggingLevel = false,
 					}));
 
             foreach (var aspProject in projects)
             {
                 var outputTarget = applicationManager.OutputTargets.Single(x => x.Id == aspProject.Id);
 
-                registry.RegisterTemplate(TemplateId, outputTarget, target => new AppSettingsTemplate(target, new AppSettingsModel(null, aspProject.Location, aspProject.RequiresSpecifiedRole, aspProject.IncludeAllowHosts)));
+                registry.RegisterTemplate(TemplateId, outputTarget, target => new AppSettingsTemplate(target, new AppSettingsModel(null, aspProject.Location, aspProject.RequiresSpecifiedRole, aspProject.IncludeAllowHosts, aspProject.IncludeAspNetCoreLoggingLevel)));
 
                 foreach (var runtimeEnvironment in aspProject.RuntimeEnvironments)
                 {
-                    registry.RegisterTemplate(TemplateId, outputTarget, target => new AppSettingsTemplate(target, new AppSettingsModel(runtimeEnvironment, aspProject.Location, aspProject.RequiresSpecifiedRole, aspProject.IncludeAllowHosts)));
+                    registry.RegisterTemplate(TemplateId, outputTarget, target => new AppSettingsTemplate(target, new AppSettingsModel(runtimeEnvironment, aspProject.Location, aspProject.RequiresSpecifiedRole, aspProject.IncludeAllowHosts, aspProject.IncludeAspNetCoreLoggingLevel)));
                 }
             }
         }